1_project_data_modeling_with_postgres/requirements.txt
pandas==1.2.1
psycopg2-binary==2.8.6